refactor(phase-19)(pr-4b): ts 파일 분할 (editor/api + GameChat + FriendsList)
F-react-3·4 해소 — TS·TSX 400줄 한도 초과 3파일을 디렉터리 + 배럴 구조로 분할.

editor/api.ts (428줄) → features/editor/api/ 배럴 디렉터리
  - types.ts (191), keys.ts (20), themes.ts (97), characters.ts (68),
    content.ts (51), moduleSchemas.ts (16), index.ts (88)
  - index.ts는 type-only + 명시 named re-export (tree-shake 유지)
  - editorClueApi / editorMapApi는 기존 대로 barrel에서 재-export

GameChat.tsx (423줄) → features/game/components/GameChat/ 디렉터리
  - types.ts (74): 페이로드·tab·상수
  - useGameChatState.ts (186): 상태·WS 이벤트 구독 훅
  - TabHeader (50), WhisperTargetSelect (29),
    GroupSidebar (50) + GroupListView (53) + GroupCreateView (64),
    MessageList (40), MessageComposer (49)
  - index.tsx (64): 컴포넌트 합성만 담당

FriendsList.tsx (415줄) → features/social/components/FriendsList/ 디렉터리
  - FriendRow (80), PendingRow (61), AddFriendModal (56)
  - TabSwitcher (43), FriendsTab (83), PendingTab (49)
  - index.tsx (121): 상태·뮤테이션 소유 + 탭별 서브컴포넌트 렌더

호출부 import 경로는 불변 — @/features/editor/api, @/features/social/components,
@/features/game/components/GameChat 모두 그대로 동작.

검증:
  - tsc --noEmit: 0 errors
  - eslint: 0 errors (41 pre-existing warnings unchanged)
  - vitest: 109 files / 1047 tests passed
  - vite build: success
