+++ apps/web/src/features/editor/api/keys.ts
@@ -1,0 +1,20 @@
+// ---------------------------------------------------------------------------
+// Editor API — React Query keys
+// ---------------------------------------------------------------------------
+
+export const editorKeys = {
+  all: ["editor"] as const,
+  themes: () => [...editorKeys.all, "themes"] as const,
+  theme: (id: string) => [...editorKeys.all, "themes", id] as const,
+  characters: (themeId: string) =>
+    [...editorKeys.all, "themes", themeId, "characters"] as const,
+  maps: (themeId: string) =>
+    [...editorKeys.all, "themes", themeId, "maps"] as const,
+  locations: (themeId: string) =>
+    [...editorKeys.all, "themes", themeId, "locations"] as const,
+  clues: (themeId: string) =>
+    [...editorKeys.all, "themes", themeId, "clues"] as const,
+  content: (themeId: string, key: string) =>
+    [...editorKeys.all, "themes", themeId, "content", key] as const,
+  moduleSchemas: () => [...editorKeys.all, "module-schemas"] as const,
+};
